+++ lib/kariyon/periodic_creator.rb
@@ -5,19 +5,34 @@ require 'fileutils'
 module Kariyon
   class PeriodicCreator
     def self.clean
+      begin
+        raise 'MINCをアンインストールしてください。' if Kariyon::Deployer.minc?
+      rescue => e
+        puts "#{e.class}: #{e.message}"
+        exit 1
+      end
+
       Dir.glob(File.join(destroot, '*')) do |f|
-        next unless File.symlink?(f)
-        if File.readlink(f).match(ROOT_DIR)
-          puts "delete #{f}"
-          File.unlink(f)
+        begin
+          if File.symlink?(f) && File.readlink(f).match(ROOT_DIR)
+            puts "delete #{f}"
+            File.unlink(f)
+          end
+        rescue => e
+          puts "#{e.class} #{e.message}"
         end
       end
     end
 
     def self.create
-      raise 'MINCをアンインストールしてください。' if Kariyon::Deployer.minc?
-      puts "link #{src} -> #{dest}"
-      File.symlink(src, dest)
+      begin
+        raise 'MINCをアンインストールしてください。' if Kariyon::Deployer.minc?
+        puts "link #{src} -> #{dest}"
+        File.symlink(src, dest)
+      rescue => e
+        puts "#{e.class}: #{e.message}"
+        exit 1
+      end
     end
 
     private
